+++ lib/widgets/calendar/calendar_indicators.dart
@@ -37,7 +37,7 @@ class CheckMark extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return const Text(
+    return Text(
       '✓',
       style: TextStyle(
         fontSize: 9,