+++ lib/widgets/events/color_picker.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 import '../../theme/app_colors.dart';
 
-const eventPalette = <Color>[
+final eventPalette = <Color>[
   AppColors.primary,
   AppColors.mint,
   AppColors.pink,